+++ project.go
@@ -49,6 +49,9 @@ type Project struct {
 	ID   string `jsonapi:"primary,projects"`
 	Name string `jsonapi:"attr,name"`
 
+	// **Note: This field is still in BETA and subject to change.**
+	Description string `jsonapi:"attr,description"`
+
 	// Relations
 	Organization *Organization `jsonapi:"relation,organization"`
 }
@@ -76,6 +79,10 @@ type ProjectCreateOptions struct {
 
 	// Required: A name to identify the project.
 	Name string `jsonapi:"attr,name"`
+
+	// Optional: A description for the project.
+	// **Note: This field is still in BETA and subject to change.**
+	Description *string `jsonapi:"attr,description,omitempty"`
 }
 
 // ProjectUpdateOptions represents the options for updating a project
@@ -88,6 +95,10 @@ type ProjectUpdateOptions struct {
 
 	// Optional: A name to identify the project
 	Name *string `jsonapi:"attr,name,omitempty"`
+
+	// Optional: A description for the project.
+	// **Note: This field is still in BETA and subject to change.**
+	Description *string `jsonapi:"attr,description,omitempty"`
 }
 
 // List all projects.